client/requirements.txt
loguru
python-socketio
requests